+++ LookupAnything/Framework/Fields/Models/RecipeItemEntry.cs
@@ -1,4 +1,5 @@
 using Pathoschild.Stardew.Common;
+using StardewValley;
 
 namespace Pathoschild.Stardew.LookupAnything.Framework.Fields.Models;
 
@@ -8,4 +9,5 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Fields.Models;
 /// <param name="Quality">The item quality that will be produced, if applicable.</param>
 /// <param name="IsGoldPrice">Whether this is a gold price, rather than an ingredient.</param>
 /// <param name="IsValid">Whether this recipe input or output is valid.</param>
-internal record RecipeItemEntry(SpriteInfo? Sprite, string DisplayText, int? Quality, bool IsGoldPrice, bool IsValid = true);
+/// <param name="Entity">The output entity represented by this entity, if applicable.</param>
+internal record RecipeItemEntry(SpriteInfo? Sprite, string DisplayText, int? Quality, bool IsGoldPrice, bool IsValid = true, object? Entity = null);