+++ CentralStation/docs/release-notes.md
@@ -1,7 +1,9 @@
 ﻿[← back to readme](README.md)
 
 # Release notes
-## Upcoming release
+## 1.2.1
+Released 26 March 2025 for SMAPI 4.1.10 or later.
+
 * For players:
   * Improved error message when the content pack isn't installed to be clearer.
   * Fixed edge case where a ticket machine could disappear if a mod reloaded its location's map while you were there.